feat(editor): Restructure folders for new components
