+++ apps/metadata-editor/src/app/records/all-records/all-records-list.component.html
