+++ apps/metadata-editor/src/app/records/all-records/all-records-list.component.ts
@@ -5,7 +5,8 @@ import { Router } from '@angular/router'
 import { CatalogRecord } from '@geonetwork-ui/common/domain/record'
 import { SearchFacade } from '@geonetwork-ui/feature/search'
 import { UiSearchModule } from '@geonetwork-ui/ui/search'
-import { DashboardSearchService } from '../dashboard/dashboard-search.service'
+import { DashboardSearchService } from '../../dashboard/dashboard-search.service'
+
 @Component({
   selector: 'md-editor-all-records-list',
   templateUrl: './all-records-list.component.html',
